Diplom/my_diplom/requirements.txt
django
scikit-learn
torch
torchvision
tensorflow
psycopg2-binary
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Pillow